+++ blocks/cmp-container/cmp-container.js
@@ -1,4 +1,10 @@
+import {
+  imageClickable,
+} from '../form/util.js';
+
 export default function decorate(block) {
+  // Logo Clickable
+  imageClickable('header .cmp-container > div:nth-child(1) > div > picture > img', 'https://www.hdfcbank.com/');
   const image = block.querySelector('picture');
   if (image) {
     document.querySelector('header').append(block);
