+++ blocks/form/util.js
@@ -201,6 +201,16 @@ function updateRequiredCheckboxGroup(name, htmlForm) {
   });
 }
 
+export function imageClickable(selector, url) {
+  const element = document.querySelector(selector);
+  if (element) {
+    element.addEventListener('click', (event) => {
+      event.preventDefault();
+      window.open(url, '_blank');
+    });
+  }
+}
+
 export function checkValidation(fieldElement) {
   const wrapper = fieldElement.closest('.field-wrapper');
   const isCheckboxGroup = fieldElement.dataset.fieldType === 'checkbox-group';
